Refactor code to better use static casting and nullptr checks
This change mainly refactors parts of the code to use static_cast<uint32_t>() instead of static_cast<int>(), and also improves the code format in several places. In addition, it replaces the traditional C-style null pointer check by using nullptr instead. The commit further makes some changes in how code resources were being handled and ensures that padding is properly initialized to be zero.

+++ src/ByteCodeGen_Cast.cpp
@@ -788,7 +788,7 @@ bool ByteCodeGen::emitCast(ByteCodeGenContext* context, AstNode* exprNode, TypeI
         }
         else
         {
-            SWAG_ASSERT(exprNode->resultRegisterRc.size() == static_cast<int>(typeInfo->numRegisters()));
+            SWAG_ASSERT(exprNode->resultRegisterRc.size() == typeInfo->numRegisters());
             node->resultRegisterRc = exprNode->resultRegisterRc;
         }
